[F] All Links, [A] Footer, Construction
[F] - [Fix]
Contribution links are now defined within the footer.
It takes the Base URL from the _config.yml !

And footer itself is defined in site_nav

[A] - [ADD]
Footer.html
(Has the link for contribution informationI).

Underconstruction.html
(rooting for shortened link when data of the page is missing.

Organized pages to their respective positions and categories.