[ADD] Back To Top, Edit Date
An option to scroll back to the top, to avoid doom scrolling manually on longer Content pages.

Edit Date now is using a dependency Pluging, that captures the editing date of the given file.

It is installed in the GemFile, uploaded to gemfile.lock

Which is then called in config.yml as a plugin.

Which then is used through Footer.html
to ask for the given date and specify for each file.