[CHANGE] Sidebar, Footer, search
Side bar no longer shows the contributors or user or people, but instead now references an open source that is hosted at..

Search header is added so that, the Search bar is changed from the Website title to Self defined title.

Footer previously added " -#top " at the end of the header now it should instead just Reload or refresh the page that makes it go to the top.