[Change]Collection Loca, Fix warning calls
Changes the locations of the collections from custom.scss to instead be imported there, to make it easier in the future to maintain.

Which is now in " Collections.scss"

Fixed Language c++ ->
SubDomain pages

On there .warning calls were wrong and wouldn't be displayed on the Website.
It must be included as a title type above the content type.

.warning
content