Method to Store Content shown
Additionally added a feature that adds an event listener and saves the current language to the users local storage, if not defiend previously.

If it is defined it loads the selected language from this example Cpp or Csharp.

Which will be asked each time a refresh is called or page is reloaded.

Additionally Pressing the Button either cpp or Csharp will rewrite the Language Preference and store it to locally.

Note:

This Requires that the JavaScript would be made globally not for each instance where possible.

Make sure that only specified buttons are called so make cases if there are less or more languages.

" in the future figure out where to store these as all in the same file would be an overkill potentially ".