Save PContent Specific To page
Now the Content has a saved value based on the Path of the URL and the ID selected.  When page is refreshed or loaded it would load the value if defined by default it should be on " cpp "

Wrote comments to be more clear  and understandable what each section actually does.

Made an " asset " folder to hold the js file. Additionally in the next commit will include Images and CSS into that specified location.

Currently Using " OOP and Memory " as placeholders for these examples.