Organized Folder Structure
Now the following structure is based on the " Just the Docs " default theme to make it easier to keep track of changes if there is a need to.

Making it easier to gain an understanding of the changes made with reference to the Theme's own repository.

Also Modified the images to look for the new location of " Change Project Name " page and " Default " under construction image when content is missing.