Visual Update to Collections/Code
Collections are now documented and organized specific to the Light theme of the current website, using Hex Color scheme, and titles can now be specified manually for each collection by the use of : {: .collection title="NewTitle"} if it is not defined the specified collection uses the default written title which is based on the name of itself.

Note that collection Title can be written as empty as well if there is not a requirement to display it!

Just use :  title= ""  instead.

Made Adjustments within Code.scss, where each code block would not display a box-shadow around itself, as well as on the left side of it use a border colour.