Removal of Under Construction Sign
As there is now a specified Warning message, I think the Image is no longer required, and it would as well free up resources, to be less rate-limited in the future (The less Images the better).

I think it also changes the Visual appeal of the website to be more lack-luster if left on, with that decision it has now been Removed.