Dark and Light switch
With the current modification the theme selected is stored locally with an eventListener that stores that information and asks for it.

By default the website will be on a white theme, unless the user changes it, in theory the selection should stay fixed even when the user leave the site and comes back later.

For the icons I used a the default windows emojis of the Sun and moon, which I think work great and show it off to its best abilities.

Additionally I left the Border on the icon so indicate that it can be selected or changed.

Doesn't seem to have a performance issues.

Does have an issue that page by default is white themed, so upon changing it flickers with the lights, occasionally on slow internets it may show behind the scenes of the website which could be an issue but that is up for debate.

See files changed

config -> Selects the default color scheme
Aux_nav -> adds a new nav list item to the top right panel where " Contribute to .. " is located as a label button

Head.html -> Modifies the stylesheets and adds a script of the default " just the docs " theme .js  Essentially it has been fixated to the original code

Default.html -> Adds our local script " colorThemeSelection.js " this is what controls the theme button and selection and load.

ColorThemeSelection.js -> As described